services/rag-chat/requirements.txt
# Required for RAGEngine text splitting
langchain_text_splitters
fastapi==0.109.0
uvicorn[standard]==0.27.0
langchain-community
faiss-cpu==1.13.1
python-dotenv==1.0.0
redis==5.0.1
boto3==1.34.0
aiohttp==3.9.1
pydantic==2.12.5
pydantic-settings==2.12.0
aiofiles==23.2.1
langchain==1.2.0
openai==2.14.0
tiktoken==0.12.0
flower==2.0.1
langchain-openai==1.1.6
weaviate-client==3.26.7
langchain-classic>=1.0.0